requirements.txt
# Django and web framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Computer vision and ML
opencv-python==4.8.1.78
deepface==0.0.79
face-recognition==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
dlib==19.24.2
cmake==3.27.2

# Database and utilities
Pillow==10.0.1
python-decouple==3.8
psycopg2-binary==2.9.7  # If using PostgreSQL
whitenoise==6.6.0

# Development and debugging
django-debug-toolbar==4.2.0
python-dotenv==1.0.0